NAS-114326: Refactor Jobs related components
